Fixed scenario names
